fix: project structure
