+++ src/components/header/contacts.scss
@@ -1,0 +1,55 @@
+.contacts {
+  background-color: $bg-color-black;
+  color: $color-white;
+  font-size: 12px;
+  line-height: 18px;
+
+  &__wrap {
+    display: flex;
+    justify-content: space-between;
+  }
+  
+  &__text {
+    color: $text-color-light;
+  }
+  
+  &__link {
+    color: $color-white;
+    padding-left: 5px;
+    transition: 0.3s;
+    text-decoration: none;
+    padding-right: 3px;
+
+    &:hover {
+      opacity: 0.6;
+      transition: 0.3s;
+    }
+
+    &::before {
+      content: "Call Us:";
+      padding-right: 5px;
+    }
+  }
+
+  &__communication {
+    display: flex;
+    align-items: center;
+    box-sizing: content-box;
+    width: content-box;
+  }
+
+  &__messenger-link {
+    padding-left: 13px;
+  }
+
+  &__messenger-icon {
+    display: block;
+    transition: 0.3s;
+
+    &:hover {
+      transform: scale(1.15);
+      opacity: 0.6;
+      transition: 0.3s;
+    }
+  }
+}
