+++ src/index.html
@@ -10,65 +10,65 @@
 </head>
 <body class="body">
   <header class="header">
-    <div class="header__contacts">
-      <div class="container header__wrap">
+    <section class="contacts">
+      <div class="container contacts__wrap">
         <p>
-          <span class="header__wrap_text">Mon-Thu:</span> 9:00 AM - 5:30 PM
+          <span class="contacts__text">Mon-Thu:</span> 9:00 AM - 5:30 PM
         </p>
         <p>
-          <span class="header__wrap_text">Visit our showroom in 1234 Street Address City Address, 1234</span>
-          <a href="#" class="header__wrap_link">Contact Us</a>
+          <span class="contacts__text">Visit our showroom in 1234 Street Address City Address, 1234</span>
+          <a href="#" class="contacts__link">Contact Us</a>
         </p>
-        <div class="header__wrap_communication">
-          <a href="#" class="header__wrap_link">(00) 1234 5678</a>
-          <a href="https://www.facebook.com/" target="_blank" rel="noopener" class="messenger-link">
-            <img src="assets/icons/facebook-icon.svg" alt="Facebook" title="Facebook" class="messenger-link_icon">
+        <div class="contacts__communication">
+          <a href="#" class="contacts__link">(00) 1234 5678</a>
+          <a href="https://www.facebook.com/" target="_blank" rel="noopener" class="contacts__messenger-link">
+            <img src="assets/icons/facebook-icon.svg" alt="Facebook" title="Facebook" class="contacts__messenger-icon">
           </a>
-          <a href="https://www.instagram.com/" target="_blank" rel="noopener" class="messenger-link">
-            <img src="assets/icons/instagram-icon.svg" alt="Instagram" title="Instagram" class="messenger-link_icon">
+          <a href="https://www.instagram.com/" target="_blank" rel="noopener" class="contacts__messenger-link">
+            <img src="assets/icons/instagram-icon.svg" alt="Instagram" title="Instagram" class="contacts__messenger-icon">
           </a>
         </div>
       </div>
-    </div>
+    </section>
     <div class="container">
-      <div class="header__wrap nav__wrap">
+      <div class="header__wrap">
         <a href="#" class="header__logo">
           <img src="assets/icons/logo.svg" alt="logo">
         </a>
-        <nav class="header__nav">
-          <ul class="header__nav_list">
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Laptops</a>
+        <nav class="nav-menu">
+          <ul class="nav-menu__list">
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Laptops</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Smartphones</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Smartphones</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Fragrances</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Fragrances</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Skincare</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Skincare</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Groceries</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Groceries</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Home-decoration</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Home-decoration</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Furniture</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Furniture</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Tops</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Tops</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Dresses</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Dresses</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Shoes</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Shoes</a>
             </li>
-            <li class="header__nav_item">
-              <a href="#" class="item_link">Shirts</a>
+            <li class="nav-menu__item">
+              <a href="#" class="nav-menu__item-link">Shirts</a>
             </li>
           </ul>
         </nav>
@@ -79,7 +79,7 @@
           </div>
           <div class="header__basket">
             <img src="assets/icons/basket-icon.svg" alt="basket" title="basket">
-            <div class="header__basket_counter">0</div>
+            <div class="header__basket-counter">0</div>
           </div>
           <div class="header__user">
             <img src="assets/icons/user-icon.svg" alt="user-icon">
@@ -91,9 +91,9 @@
   <main class="main">
     <div class="container">
       <div class="main__wrap">
-        <aside class="side-bar">
+        <section class="sort-bar">
           filter options
-        </aside>
+        </section>
         <section class="catalog">
           catalog 
         </section>
