+++ src/style.scss
@@ -2,3 +2,9 @@
 @import 'styles/font';
 @import 'styles/variables';
 @import 'styles/main';
+
+// components
+@import 'components/header/header.scss';
+@import 'components/header/contacts.scss';
+@import 'components/header/nav-menu.scss';
+@import 'components/footer/footer.scss';
